refactor: split arf-harp completion.rs by tier
completion.rs mixed pure string analysis, filesystem scanning, and unsafe
R interop in 1176 lines. The dependencies between them already ran one
way, so the seams were there to be used.

It becomes completion/, with context detection and the paren and namespace
checks in one module, package scanning and its cache in another, and every
unsafe R call in a third, behind a thin mod.rs holding the entry point and
the stderr guard. escape_r_string sits with the R interop despite not
being unsafe itself, since generating R code is all it does.

Each test module moves next to the code it covers. The public API is
unchanged, so no caller outside the crate is touched, and the test count
is unchanged at 1149.
